+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 0.0.2
 
-  * ...
+  * Fix issue 1: OpenURI returns a StringIO not TempFile when size is under 10KB
 
 ## Version 0.0.1